+++ src/Data/Geo/Jord/HorizontalPosition.hs
@@ -6,17 +6,17 @@
 -- Stability:   experimental
 -- Portability: portable
 --
--- Types and functions for working with positions.
+-- Types and functions for working with horizontal positions.
 --
 -- All functions are implemented using the vector-based approached described in
 -- <http://www.navlab.net/Publications/A_Nonsingular_Horizontal_Position_Representation.pdf Gade, K. (2010). A Non-singular Horizontal Position Representation>
 --
 -- This module assumes a spherical earth.
 --
-module Data.Geo.Jord.Position
+module Data.Geo.Jord.HorizontalPosition
     (
-    -- * The 'Position' type
-      Position(..)
+    -- * The 'HorizontalPosition' type
+      HorizontalPosition(..)
     -- * Geodetic calculations
     , angularDistance
     , antipode
@@ -38,28 +38,28 @@ module Data.Geo.Jord.Position
 import Data.Geo.Jord.Angle
 import Data.Geo.Jord.LatLong
 import Data.Geo.Jord.Length
-import Data.Geo.Jord.NVector
 import Data.Geo.Jord.Quantity
+import Data.Geo.Jord.Vector3d
 import Data.List (subsequences)
 import Prelude hiding (fail)
 
--- | The 'Position' class defines 2 functions to convert a position to and from a 'NVector'.
--- All functions in this module first convert 'Position' to 'NVector' and any resulting 'NVector' back
--- to a 'Position'. This allows the call site to pass either 'NVector' or 'LatLong' and to get back
--- the same class instance.
-class Position a where
-    -- | Converts a 'NVector' into 'Position' instance.
-    fromNVector :: NVector -> a
-    -- | Converts the 'Position' instance into a 'NVector'.
-    toNVector :: a -> NVector
+-- | The 'HorizontalPosition' class defines 2 functions to convert a Horizontal Position to and from a n-vector.
+-- All functions in this module first convert 'HorizontalPosition' to a n-vector and any resulting n-vector back
+-- to a 'HorizontalPosition'. This allows the call site to pass either n-vector or another 'HorizontalPosition' instance
+-- and to get back the same class instance.
+class (Eq a) => HorizontalPosition a where
+    -- | Converts a 'Vector3d' into 'HorizontalPosition' instance.
+    fromNVector :: Vector3d -> a
+    -- | Converts the 'HorizontalPosition' instance into a 'Vector3d'.
+    toNVector :: a -> Vector3d
 
--- | 'LatLong' to/from 'NVector'.
-instance Position LatLong where
+-- | 'LatLong' to/from 'Vector3d'.
+instance HorizontalPosition LatLong where
     fromNVector v = latLong lat lon
       where
         lat = atan2' (z v) (sqrt (x v * x v + y v * y v))
         lon = atan2' (y v) (x v)
-    toNVector g = nvector x' y' z'
+    toNVector g = Vector3d x' y' z'
       where
         lat = latitude g
         lon = longitude g
@@ -69,34 +69,40 @@ instance Position LatLong where
         z' = sin' lat
 
 -- | Identity.
-instance Position NVector where
+instance HorizontalPosition Vector3d where
     fromNVector v = v
     toNVector v = v
 
--- | Angle between the two given 'NVector's.
+-- | Angle between the two given 'HorizontalPosition's.
+-- If @n@ is 'Nothing', the angle is always in [0..180], otherwise it is in [-180, +180],
+-- signed + if @p1@ is clockwise looking along @n@, - in opposite direction.
+angularDistance :: (HorizontalPosition a) => a -> a -> Maybe a -> Angle
+angularDistance p1 p2 n = angularDistance' (toNVector p1) (toNVector p2) (fmap toNVector n)
+
+-- | Angle between the two given 'Vector3d's.
 -- If @n@ is 'Nothing', the angle is always in [0..180], otherwise it is in [-180, +180],
 -- signed + if @v1@ is clockwise looking along @n@, - in opposite direction.
-angularDistance :: NVector -> NVector -> Maybe NVector -> Angle
-angularDistance v1 v2 n = atan2' sinO cosO
+angularDistance' :: Vector3d -> Vector3d -> Maybe Vector3d -> Angle
+angularDistance' v1 v2 n = atan2' sinO cosO
   where
     sign = maybe 1 (signum . dot (cross v1 v2)) n
     sinO = sign * norm (cross v1 v2)
     cosO = dot v1 v2
 
--- | Returns the antipodal 'Position' of the given 'Position' - i.e. the position on the surface
--- of the Earth which is diametrically opposite to the given position.
-antipode :: (Position a) => a -> a
+-- | Returns the antipodal 'HorizontalPosition' of the given 'HorizontalPosition' - i.e. the HorizontalPosition on the surface
+-- of the Earth which is diametrically opposite to the given HorizontalPosition.
+antipode :: (HorizontalPosition a) => a -> a
 antipode p = fromNVector (scale (toNVector p) (-1.0))
 
 -- | 'destination'' assuming a radius of 'meanEarthRadius'.
-destination :: (Position a) => a -> Angle -> Length -> a
+destination :: (HorizontalPosition a) => a -> Angle -> Length -> a
 destination p b d = destination' p b d meanEarthRadius
 
--- | Computes the destination 'Position' from the given 'Position' having travelled the given distance on the
+-- | Computes the destination 'HorizontalPosition' from the given 'HorizontalPosition' having travelled the given distance on the
 -- given initial bearing (bearing will normally vary before destination is reached) and using the given earth radius.
 --
 -- This is known as the direct geodetic problem.
-destination' :: (Position a) => a -> Angle -> Length -> Length -> a
+destination' :: (HorizontalPosition a) => a -> Angle -> Length -> Length -> a
 destination' p b d r
     | isZero d = p
     | otherwise = fromNVector (add (scale v (cos' ta)) (scale de (sin' ta)))
@@ -108,34 +114,31 @@ destination' p b d r
     de = add (scale nd (cos' b)) (scale ed (sin' b)) -- unit vector in the direction of the azimuth
 
 -- | 'distance'' assuming a radius of 'meanEarthRadius'.
-distance :: (Position a) => a -> a -> Length
+distance :: (HorizontalPosition a) => a -> a -> Length
 distance p1 p2 = distance' p1 p2 meanEarthRadius
 
 -- | Computes the surface distance (length of geodesic) in 'Meters' assuming a
--- spherical Earth between the two given 'Position's and using the given earth radius.
-distance' :: (Position a) => a -> a -> Length -> Length
-distance' p1 p2 = arcLength (angularDistance v1 v2 Nothing)
-  where
-    v1 = toNVector p1
-    v2 = toNVector p2
+-- spherical Earth between the two given 'HorizontalPosition's and using the given earth radius.
+distance' :: (HorizontalPosition a) => a -> a -> Length -> Length
+distance' p1 p2 = arcLength (angularDistance p1 p2 Nothing)
 
--- | Computes the final bearing arriving at given destination  @p2@ 'Position' from given 'Position' @p1@.
+-- | Computes the final bearing arriving at given destination  @p2@ 'HorizontalPosition' from given 'HorizontalPosition' @p1@.
 --  the final bearing will differ from the 'initialBearing' by varying degrees according to distance and latitude.
--- Returns 180 if both position are equals.
-finalBearing :: (Position a) => a -> a -> Angle
+-- Returns 180 if both HorizontalPosition are equals.
+finalBearing :: (HorizontalPosition a) => a -> a -> Angle
 finalBearing p1 p2 = normalise (initialBearing p2 p1) (decimalDegrees 180)
 
--- | Computes the initial bearing from given @p1@ 'Position' to given @p2@ 'Position', in compass degrees.
--- Returns 0 if both position are equals.
-initialBearing :: (Position a) => a -> a -> Angle
-initialBearing p1 p2 = normalise (angularDistance gc1 gc2 (Just v1)) (decimalDegrees 360)
+-- | Computes the initial bearing from given @p1@ 'HorizontalPosition' to given @p2@ 'HorizontalPosition', in compass degrees.
+-- Returns 0 if both HorizontalPosition are equals.
+initialBearing :: (HorizontalPosition a) => a -> a -> Angle
+initialBearing p1 p2 = normalise (angularDistance' gc1 gc2 (Just v1)) (decimalDegrees 360)
   where
     v1 = toNVector p1
     v2 = toNVector p2
     gc1 = cross v1 v2 -- great circle through p1 & p2
     gc2 = cross v1 northPole -- great circle through p1 & north pole
 
--- | Computes the 'Position' at given fraction @f@ between the two given 'Position's @p0@ and @p1@.
+-- | Computes the 'HorizontalPosition' at given fraction @f@ between the two given 'HorizontalPosition's @p0@ and @p1@.
 --
 -- Special conditions:
 --
@@ -146,7 +149,7 @@ initialBearing p1 p2 = normalise (angularDistance gc1 gc2 (Just v1)) (decimalDeg
 --
 -- 'error's if @f < 0 || f > 1.0@
 --
-interpolate :: (Position a) => a -> a -> Double -> a
+interpolate :: (HorizontalPosition a) => a -> a -> Double -> a
 interpolate p0 p1 f
     | f < 0 || f > 1 = error ("fraction must be in range [0..1], was " ++ show f)
     | f == 0 = p0
@@ -156,15 +159,15 @@ interpolate p0 p1 f
     v0 = toNVector p0
     v1 = toNVector p1
 
--- | Determines whether the given 'Position' is inside the polygon defined by the given list of 'Position's.
+-- | Determines whether the given 'HorizontalPosition' is inside the polygon defined by the given list of 'HorizontalPosition's.
 -- The polygon is closed if needed (i.e. if @head ps /= last ps@).
 --
 -- Uses the angle summation test: on a sphere, due to spherical excess, enclosed point angles
 -- will sum to less than 360°, and exterior point angles will be small but non-zero.
 --
--- Always returns 'False' if positions does not at least defines a triangle.
+-- Always returns 'False' if HorizontalPositions does not at least defines a triangle.
 --
-isInside :: (Eq a, Position a) => a -> [a] -> Bool
+isInside :: (Eq a, HorizontalPosition a) => a -> [a] -> Bool
 isInside p ps
     | null ps = False
     | head ps == last ps = isInside p (init ps)
@@ -177,14 +180,14 @@ isInside p ps
     es = egdes (map (sub v . toNVector) ps)
 
 -- | [p1, p2, p3, p4] to [(p1, p2), (p2, p3), (p3, p4), (p4, p1)]
-egdes :: [NVector] -> [(NVector, NVector)]
+egdes :: [Vector3d] -> [(Vector3d, Vector3d)]
 egdes ps = zip ps ps'
   where
     ps' = tail ps ++ [head ps]
 
--- | Computes the geographic mean 'Position' of the given 'Position's if it is defined.
+-- | Computes the geographic mean 'HorizontalPosition' of the given 'HorizontalPosition's if it is defined.
 --
--- The geographic mean is not defined for the antipodals positions (since they
+-- The geographic mean is not defined for the antipodals HorizontalPositions (since they
 -- cancel each other).
 --
 -- Special conditions:
@@ -196,7 +199,7 @@ egdes ps = zip ps ps'
 --     mean [p1, .., antipode p1] == Nothing
 -- @
 --
-mean :: (Position a) => [a] -> Maybe a
+mean :: (HorizontalPosition a) => [a] -> Maybe a
 mean [] = Nothing
 mean [p] = Just p
 mean ps =
@@ -215,10 +218,10 @@ mean ps =
 meanEarthRadius :: Length
 meanEarthRadius = metres 6371008.8
 
--- | 'Position' of the North Pole.
-northPole :: (Position a) => a
-northPole = fromNVector (nvector 0.0 0.0 1.0)
+-- | 'HorizontalPosition' of the North Pole.
+northPole :: (HorizontalPosition a) => a
+northPole = fromNVector (Vector3d 0.0 0.0 1.0)
 
--- | 'Position' of the South Pole.
-southPole :: (Position a) => a
-southPole = fromNVector (nvector 0.0 0.0 (-1.0))
+-- | 'HorizontalPosition' of the South Pole.
+southPole :: (HorizontalPosition a) => a
+southPole = fromNVector (Vector3d 0.0 0.0 (-1.0))